+++ src/main/kotlin/com/personio/synthetics/step/ui/ActionsStep.kt
@@ -3,11 +3,9 @@ package com.personio.synthetics.step.ui
 import com.datadog.api.client.v1.model.SyntheticsStep
 import com.datadog.api.client.v1.model.SyntheticsStepType
 import com.personio.synthetics.client.BrowserTest
-import com.personio.synthetics.config.isDatadogVariable
 import com.personio.synthetics.model.actions.ActionsParams
 import com.personio.synthetics.step.addStep
 import com.personio.synthetics.step.ui.model.TargetElement
-import java.net.URL
 
 private const val DEFAULT_TEXT_DELAY: Long = 25 // in milliseconds
 
@@ -59,7 +57,7 @@ fun BrowserTest.clickStep(
  * Adds a new navigation step for following a link to the synthetic browser test
  * @param stepName Name of the step
  * @param url The navigation url. You can pass url like the following
- * - only the location (eg: /test/page) for appending to the base url of the test
+ * - only the location (eg: /test/page) for staying on the same domain
  * - pass full url including http(s)://
  * - global or local variable. For using those, use the function "fromVariable(variableName)" in the parameter
  *   for example /test/page/${fromVariable("TEST")}
@@ -72,15 +70,6 @@ fun BrowserTest.navigateStep(
     f: (SyntheticsStep.() -> Unit)? = null,
 ) = addStep(stepName) {
     type = SyntheticsStepType.GO_TO_URL
-    val target =
-        if (url.isDatadogVariable()) {
-            url
-        } else {
-            runCatching { URL(url) }
-                .recover { URL(config.request.url + url) }
-                .getOrThrow()
-                .toString()
-        }
-    params = ActionsParams(value = target)
+    params = ActionsParams(value = url)
     if (f != null) f()
 }